Clean up client organization
